Fixed a bug and added some comments
Fixed a bug which was introduced in a recent edit, that completely
broke the exp-eval function.

Also added comments to exp-easy-eval, and the function assoc lists, and
removed the empty assoc lists from the end of the file.

Moreover, corrected a typo in the comment format template at the
beginning of each file which forgot one semi-colon before the
‘Example-usage’ section of the function comment format.

TODO: Make that correction in all files.